backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
beanie>=1.26.0,<2.0.0
python-dotenv>=1.0.0
openai>=1.30.0
reportlab>=4.0.0
python-multipart>=0.0.9
pydantic[email]>=2.5.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
fastapi-mail>=1.4.1
gunicorn>=21.2.0
